+++ .eslintrc
@@ -50,7 +50,7 @@
     ],
     "max-len": [
       "warn",
-      140
+      150
     ],
     "no-console": [
       "warn"
