+++ src/devices.ts
@@ -1,0 +1,222 @@
+import { Service, PlatformAccessory, CharacteristicValue } from 'homebridge';
+import { Meross } from './platform';
+import { interval, Subject } from 'rxjs';
+import { debounceTime, skipWhile, tap } from 'rxjs/operators';
+import { DevicesConfig } from './settings';
+import { AxiosResponse } from 'axios';
+
+export class Outlet {
+  private service: Service;
+
+  On!: CharacteristicValue;
+  OutletInUse!: CharacteristicValue;
+
+  plugUpdateInProgress!: boolean;
+  doPlugUpdate!: Subject<unknown>;
+  deviceStatus: any;
+
+  constructor(
+    private readonly platform: Meross,
+    private accessory: PlatformAccessory,
+    public device: DevicesConfig,
+  ) {
+    // default placeholders
+    this.On = false;
+    this.OutletInUse = true;
+
+    // this is subject we use to track when we need to POST changes to the SwitchBot API
+    this.doPlugUpdate = new Subject();
+    this.plugUpdateInProgress = false;
+
+    // Retrieve initial values and updateHomekit
+    this.refreshStatus();
+
+    // set accessory information
+    accessory
+      .getService(this.platform.Service.AccessoryInformation)!
+      .setCharacteristic(this.platform.Characteristic.Manufacturer, 'SwitchBot')
+      .setCharacteristic(this.platform.Characteristic.Model, 'SWITCHBOT-PLUG-SP11')
+      .setCharacteristic(this.platform.Characteristic.SerialNumber, device.serialNumber!)
+      .setCharacteristic(this.platform.Characteristic.FirmwareRevision, device.firmwareRevision!);
+
+    // get the WindowCovering service if it exists, otherwise create a new WindowCovering service
+    // you can create multiple services for each accessory
+    (this.service =
+      accessory.getService(this.platform.Service.Outlet) ||
+      accessory.addService(this.platform.Service.Outlet)), '%s %s', device.name!;
+
+    // To avoid "Cannot add a Service with the same UUID another Service without also defining a unique 'subtype' property." error,
+    // when creating multiple services of the same type, you need to use the following syntax to specify a name and subtype id:
+    // accessory.getService('NAME') ?? accessory.addService(this.platform.Service.Lightbulb, 'NAME', 'USER_DEFINED_SUBTYPE');
+
+    // set the service name, this is what is displayed as the default name on the Home app
+    // in this example we are using the name we stored in the `accessory.context` in the `discoverDevices` method.
+    this.service.setCharacteristic(this.platform.Characteristic.Name, device.name!);
+
+    // each service must implement at-minimum the "required characteristics" for the given service type
+    // see https://developers.homebridge.io/#/service/WindowCovering
+
+    // create handlers for required characteristics
+    this.service.getCharacteristic(this.platform.Characteristic.On).onSet(this.OnSet.bind(this));
+
+    this.service.setCharacteristic(this.platform.Characteristic.OutletInUse, this.OutletInUse || true);
+
+    // Update Homekit
+    this.updateHomeKitCharacteristics();
+
+    // Start an update interval
+    interval(this.platform.config.refreshRate! * 1000)
+      .pipe(skipWhile(() => this.plugUpdateInProgress))
+      .subscribe(() => {
+        this.refreshStatus();
+      });
+
+
+    // Watch for Plug change events
+    // We put in a debounce of 100ms so we don't make duplicate calls
+    this.doPlugUpdate
+      .pipe(
+        tap(() => {
+          this.plugUpdateInProgress = true;
+        }),
+        debounceTime(this.platform.config.pushRate! * 1000),
+      )
+      .subscribe(async () => {
+        try {
+          await this.pushChanges();
+        } catch (e) {
+          this.platform.log.error(JSON.stringify(e.message));
+          this.platform.log.debug('Plug %s -', accessory.displayName, JSON.stringify(e));
+          this.apiError(e);
+        }
+        this.plugUpdateInProgress = false;
+      });
+  }
+
+  parseStatus() {
+    this.platform.log.debug(
+      'Plug %s On: %s',
+      this.accessory.displayName,
+      this.On,
+    );
+  }
+
+  async refreshStatus() {
+    try {
+      this.platform.log.debug('Plug - Reading', `${this.platform.config.device.deviceUrl}/status`);
+      const deviceStatus = (
+        await this.platform.axios.get(`http://${this.platform.config.device.deviceUrl}/config`)
+      ).data;
+      if (deviceStatus.message === 'success') {
+        this.deviceStatus = deviceStatus;
+        this.platform.log.warn(
+          'Plug %s refreshStatus -',
+          this.accessory.displayName,
+          JSON.stringify(this.deviceStatus),
+        );
+        this.parseStatus();
+        this.updateHomeKitCharacteristics();
+      }
+    } catch (e) {
+      this.platform.log.error(
+        `Plug - Failed to refresh status of ${this.device.name}`,
+        JSON.stringify(e.message),
+        this.platform.log.debug('Plug %s -', this.accessory.displayName, JSON.stringify(e)),
+      );
+      this.apiError(e);
+    }
+  }
+
+  /**
+ * Pushes the requested changes to the SwitchBot API
+ * deviceType	commandType	  Command	    command parameter	  Description
+ * Plug   -    "command"     "turnOff"   "default"	  =        set to OFF state
+ * Plug   -    "command"     "turnOn"    "default"	  =        set to ON state
+ */
+  async pushChanges() {
+    const payload = {
+      commandType: 'command',
+      parameter: 'default',
+    } as any;
+
+    if (this.On) {
+      payload.command = 'turnOn';
+    } else {
+      payload.command = 'turnOff';
+    }
+
+    this.platform.log.info(
+      'Sending request for',
+      this.accessory.displayName,
+      'to SwitchBot API. command:',
+      payload.command,
+      'parameter:',
+      payload.parameter,
+      'commandType:',
+      payload.commandType,
+    );
+    this.platform.log.debug('Plug %s pushChanges -', this.accessory.displayName, JSON.stringify(payload));
+
+    // Make the API request
+    const push = await this.platform.axios.post(`http://${this.platform.config.device.deviceUrl}/config`, payload);
+    this.platform.log.debug('Plug %s Changes pushed -', this.accessory.displayName, push.data);
+    this.statusCode(push);
+  }
+
+  updateHomeKitCharacteristics() {
+    if (this.On !== undefined) {
+      this.service.updateCharacteristic(this.platform.Characteristic.On, this.On);
+    }
+    if (this.OutletInUse !== undefined) {
+      this.service.updateCharacteristic(this.platform.Characteristic.OutletInUse, this.OutletInUse);
+    }
+  }
+
+  public apiError(e: any) {
+    this.service.updateCharacteristic(this.platform.Characteristic.On, e);
+    this.service.updateCharacteristic(this.platform.Characteristic.OutletInUse, e);
+  }
+
+
+  private statusCode(push: AxiosResponse<any>) {
+    switch (push.data.statusCode) {
+      case 151:
+        this.platform.log.error('Command not supported by this device type.');
+        break;
+      case 152:
+        this.platform.log.error('Device not found.');
+        break;
+      case 160:
+        this.platform.log.error('Command is not supported.');
+        break;
+      case 161:
+        this.platform.log.error('Device is offline.');
+        break;
+      case 171:
+        this.platform.log.error('Hub Device is offline.');
+        break;
+      case 190:
+        this.platform.log.error('Device internal error due to device states not synchronized with server. Or command fomrat is invalid.');
+        break;
+      case 100:
+        if (this.platform.debugMode) {
+          this.platform.log.info('Command successfully sent.');
+        }
+        break;
+      default:
+        this.platform.log.debug('Unknown statusCode.');
+    }
+  }
+
+  /**
+   * Handle requests to set the value of the "Target Position" characteristic
+   */
+  OnSet(value: CharacteristicValue) {
+    this.platform.log.debug('Plug %s - Set On: %s', this.accessory.displayName, value);
+
+    this.On = value;
+    this.doPlugUpdate.next();
+  }
+
+
+}
